Refactor Thales integration to use Overtime configurations and event ABIs
