+++ options/thales/index.ts
@@ -1,7 +1,7 @@
 import { Adapter, FetchOptions, FetchResultV2 } from "../../adapters/types";
 import { addTokensReceived } from '../../helpers/token';
-import { THALES_CHAIN_CONFIG, THALES_CONTRACT_ADDRESSES } from './config';
-import { THALES_EVENT_ABI } from './abis';
+import { OVERTIME_CHAIN_CONFIG, OVERTIME_CONTRACT_ADDRESSES } from './config';
+import { OVERTIME_EVENT_ABI } from './abis';
 import { 
   parseTicketCreatedEvent, 
   parseBoughtFromAmmEvent, 
@@ -10,82 +10,106 @@ import {
 } from './parsers';
 import { CHAIN } from "../../helpers/chains";
 
-function getChainContractsToQuery(chain: string, dailyVolume: ReturnType<FetchOptions['createBalances']>) {
+function getChainContractsToQuery(
+  chain: string, 
+  dailyNotionalVolume: ReturnType<FetchOptions['createBalances']>,
+  dailyPremiumVolume: ReturnType<FetchOptions['createBalances']>
+) {
   switch(chain) {
     case CHAIN.OPTIMISM: {
       const { sportsAMMV2, thalesAMM, rangedAMM, speedMarket, chainedSpeedMarket } 
-        = THALES_CONTRACT_ADDRESSES[CHAIN.OPTIMISM];
+        = OVERTIME_CONTRACT_ADDRESSES[CHAIN.OPTIMISM];
       return [
         {
           address: sportsAMMV2,
-          eventAbi: THALES_EVENT_ABI.ticketCreated,
-          parser: (log: any) => parseTicketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.ticketCreated,
+          parser: (log: any) => parseTicketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: thalesAMM,
-          eventAbi: THALES_EVENT_ABI.boughtFromAmm,
-          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.boughtFromAmm,
+          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: rangedAMM,
-          eventAbi: THALES_EVENT_ABI.boughtFromAmm,
-          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.boughtFromAmm,
+          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: speedMarket,
-          eventAbi: THALES_EVENT_ABI.speedMarketCreated,
-          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.speedMarketCreated,
+          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: chainedSpeedMarket,
-          eventAbi: THALES_EVENT_ABI.chainedMarketCreated,
-          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.chainedMarketCreated,
+          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         }
       ];
     }
     case CHAIN.ARBITRUM: {
       const { sportsAMMV2, thalesAMM, rangedAMM, speedMarket, chainedSpeedMarket } 
-        = THALES_CONTRACT_ADDRESSES[CHAIN.ARBITRUM];
+        = OVERTIME_CONTRACT_ADDRESSES[CHAIN.ARBITRUM];
       return [
         {
           address: sportsAMMV2,
-          eventAbi: THALES_EVENT_ABI.ticketCreated,
-          parser: (log: any) => parseTicketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.ticketCreated,
+          parser: (log: any) => parseTicketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: thalesAMM,
-          eventAbi: THALES_EVENT_ABI.boughtFromAmm,
-          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.boughtFromAmm,
+          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: rangedAMM,
-          eventAbi: THALES_EVENT_ABI.boughtFromAmm,
-          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.boughtFromAmm,
+          parser: (log: any) => parseBoughtFromAmmEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: speedMarket,
-          eventAbi: THALES_EVENT_ABI.speedMarketCreated,
-          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.speedMarketCreated,
+          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: chainedSpeedMarket,
-          eventAbi: THALES_EVENT_ABI.chainedMarketCreated,
-          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.chainedMarketCreated,
+          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         }
       ];
     }
     case CHAIN.BASE: {
-      const { speedMarket, chainedSpeedMarket } = THALES_CONTRACT_ADDRESSES[CHAIN.BASE];
+      const { sportsAMMV2, speedMarket, chainedSpeedMarket } = OVERTIME_CONTRACT_ADDRESSES[CHAIN.BASE];
       return [
         {
+          address: sportsAMMV2,
+          eventAbi: OVERTIME_EVENT_ABI.ticketCreated,
+          parser: (log: any) => parseTicketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
+        },
+        {
           address: speedMarket,
-          eventAbi: THALES_EVENT_ABI.speedMarketCreated,
-          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.speedMarketCreated,
+          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         },
         {
           address: chainedSpeedMarket,
-          eventAbi: THALES_EVENT_ABI.chainedMarketCreated,
-          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyVolume)
+          eventAbi: OVERTIME_EVENT_ABI.chainedMarketCreated,
+          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
+        }
+      ];
+    }
+    case CHAIN.POLYGON: {
+      const { speedMarket, chainedSpeedMarket } = OVERTIME_CONTRACT_ADDRESSES[CHAIN.POLYGON];
+      return [
+        {
+          address: speedMarket,
+          eventAbi: OVERTIME_EVENT_ABI.speedMarketCreated,
+          parser: (log: any) => parseSpeedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
+        },
+        {
+          address: chainedSpeedMarket,
+          eventAbi: OVERTIME_EVENT_ABI.chainedMarketCreated,
+          parser: (log: any) => parseChainedMarketCreatedEvent(log, dailyNotionalVolume, dailyPremiumVolume)
         }
       ];
     }
@@ -95,8 +119,9 @@ function getChainContractsToQuery(chain: string, dailyVolume: ReturnType<FetchOp
 }
 
 export async function fetch(options: FetchOptions): Promise<FetchResultV2> {
-  const dailyVolume = options.createBalances();
-  const contractConfigs = getChainContractsToQuery(options.chain, dailyVolume);
+  const dailyNotionalVolume = options.createBalances();
+  const dailyPremiumVolume = options.createBalances();
+  const contractConfigs = getChainContractsToQuery(options.chain, dailyNotionalVolume, dailyPremiumVolume);
   
   await Promise.all(
     contractConfigs.map(async (cfg) => {
@@ -109,11 +134,13 @@ export async function fetch(options: FetchOptions): Promise<FetchResultV2> {
     })
   );
   
-  const dailyFees = await addTokensReceived({ ...THALES_CHAIN_CONFIG[options.chain], options});
+  const dailyRevenue = await addTokensReceived({ ...OVERTIME_CHAIN_CONFIG[options.chain], options });
   
   return {
-    dailyVolume,
-    dailyFees,
+    timestamp: options.startOfDay,
+    dailyNotionalVolume,
+    dailyPremiumVolume,
+    dailyRevenue,
   };
 }
 
@@ -122,15 +149,19 @@ const adapter: Adapter = {
   adapter: {
     [CHAIN.ARBITRUM]: {
       fetch,
-      start: '2024-08-01',
+      start: '2025-04-01',
     },
     [CHAIN.OPTIMISM]: {
       fetch,
-      start: '2024-08-01',
+      start: '2025-04-01',
     },
     [CHAIN.BASE]: {
       fetch,
-      start: '2024-08-01',
+      start: '2025-04-01',
+    },
+    [CHAIN.POLYGON]: {
+      fetch,
+      start: '2025-04-01',
     },
   },
 };
