+++ options/thales/parsers.ts
@@ -5,22 +5,43 @@ import {
     IChainedMarketCreatedEvent 
   } from './eventArgs';
   
-  export function parseTicketCreatedEvent(log: ITicketCreatedEvent, dailyBalance: any) {
-    const { payout, collateral } = log;
-    dailyBalance.addToken(collateral, payout);
+  export function parseTicketCreatedEvent(
+    log: ITicketCreatedEvent, 
+    dailyNotionalVolume: any,
+    dailyPremiumVolume: any
+  ) {
+    const { buyInAmount, payout, collateral } = log;
+    dailyNotionalVolume.addToken(collateral, payout);
+    dailyPremiumVolume.addToken(collateral, buyInAmount);
   }
   
-  export function parseBoughtFromAmmEvent(log: IBoughtFromAmmEvent, dailyBalance: any) {
-    const { amount } = log;
-    dailyBalance.addUSDValue(Number(amount) / 1e18);
+  export function parseBoughtFromAmmEvent(
+    log: IBoughtFromAmmEvent, 
+    dailyNotionalVolume: any,
+    dailyPremiumVolume: any
+  ) {
+    const { amount, sUSDPaid: usdcPaid } = log;
+    dailyNotionalVolume.addUSDValue(Number(amount) / 1e18);
+    dailyPremiumVolume.addUSDValue(Number(usdcPaid) / 1e6);
   }
   
-  export function parseSpeedMarketCreatedEvent(log: ISpeedMarketCreatedEvent, dailyBalance: any) {
+  export function parseSpeedMarketCreatedEvent(
+    log: ISpeedMarketCreatedEvent, 
+    dailyNotionalVolume: any,
+    dailyPremiumVolume: any
+  ) {
     const { buyinAmount } = log;
-    dailyBalance.addUSDValue(Number(buyinAmount) / 1e6); 
+    dailyNotionalVolume.addUSDValue(Number(buyinAmount) * 2 / 1e6);
+    dailyPremiumVolume.addUSDValue(Number(buyinAmount) / 1e6);
   }
   
-  export function parseChainedMarketCreatedEvent(log: IChainedMarketCreatedEvent, dailyBalance: any) {
-    const { buyinAmount, payoutMultiplier } = log;
-    dailyBalance.addUSDValue((Number(buyinAmount) / 1e6) * (Number(payoutMultiplier) / 1e18));
+  export function parseChainedMarketCreatedEvent(
+    log: IChainedMarketCreatedEvent, 
+    dailyNotionalVolume: any,
+    dailyPremiumVolume: any
+  ) {
+    const { buyinAmount, payoutMultiplier, directions } = log;
+    const notionalVolume = (Number(buyinAmount) / 1e6) * Math.pow(Number(payoutMultiplier) / 1e18, directions.length);
+    dailyNotionalVolume.addUSDValue(notionalVolume);
+    dailyPremiumVolume.addUSDValue(Number(buyinAmount) / 1e6);
   }
